requirements.txt
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Jinja2==3.1.6
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
email-validator==2.3.0
pytest==8.4.2
python-dotenv==1.1.1